requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
flask>=2.0.0
qrcode[pil]>=7.0.0
pillow>=8.0.0
